fix: add Object.values polyfill for node <=6 (#4274)
* fix: add Object.values polyfill for node <=6

* rename file

* forgot to commit eslint
